12-24 Hour time
There are now 3 functions for bot getCurrentTime and getCurrentTimeParts, when called, each will use a variable/object that (Please implement Kevin) needs to be implemented.  Based on what mode the user has set, it calls either getCurrentTime12 or getCurrentTime24 for getCurrentTime and getCurrentTimeParts12 and getCurrentTimeParts24 for getCurrentTimeParts.
So it will now return the correct time w/ am/pm based on what the user has selected (once the correct variable has been put in)
This hasn't been tested yet, Kevin you can test once you are done w/ the switch and have connected it to the function.